dev_requirements.txt
black==21.6b0
coverage==5.5
mypy==0.910
pre-commit==2.14.0
pylint==2.9.6
pylint-sqlalchemy==0.2.0
PyQt5-stubs==5.15.2.0
pytest==6.2.4
pytest-qt==4.0.2
requests==2.26.0
sqlalchemy2-stubs==0.0.2a8
types-requests==2.25.6
